Move auth runtime into client shell
Makes the composition shell own auth state, refresh scheduling, and router context while preserving the legacy auth provider as a compatibility wrapper.

Refs ENG-938
