+++ libs/client/auth/package.json
@@ -46,6 +46,7 @@
     "@shipfox/api-auth-dto": "workspace:*",
     "@shipfox/api-workspaces-dto": "workspace:*",
     "@shipfox/client-api": "workspace:*",
+    "@shipfox/client-shell": "workspace:*",
     "@shipfox/client-ui": "workspace:*",
     "@shipfox/react-ui": "workspace:*",
     "@swc/helpers": "^0.5.17",
