+++ libs/client/auth/src/components/auth-provider.tsx
@@ -1,126 +1,24 @@
-import {configureApiClient} from '@shipfox/client-api';
+import {AuthRuntime} from '@shipfox/client-shell/runtime';
 import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
-import {Provider as JotaiProvider, useAtomValue, useStore} from 'jotai';
-import {type PropsWithChildren, useEffect, useState} from 'react';
-import {getAuthRefreshDelayMs, useRefreshAuth} from '#hooks/api/refresh-auth.js';
-import {authStateAtom} from '#state/auth.js';
-
-const REFRESH_RETRY_DELAY_MS = 60_000;
+import {Provider as JotaiProvider} from 'jotai';
+import {type PropsWithChildren, useState} from 'react';
 
 export interface AuthProviderProps extends PropsWithChildren {
   queryClient?: QueryClient;
 }
 
+/**
+ * Backwards-compatible provider for applications that have not adopted
+ * `composeClientApp` yet. The shell owns the runtime behavior in both paths.
+ */
 export function AuthProvider({children, queryClient}: AuthProviderProps) {
   const [fallbackQueryClient] = useState(() => new QueryClient());
-  const client = queryClient ?? fallbackQueryClient;
 
   return (
-    <QueryClientProvider client={client}>
+    <QueryClientProvider client={queryClient ?? fallbackQueryClient}>
       <JotaiProvider>
-        <AuthProviderContent>{children}</AuthProviderContent>
+        <AuthRuntime>{children}</AuthRuntime>
       </JotaiProvider>
     </QueryClientProvider>
   );
-}
-
-function AuthProviderContent({children}: PropsWithChildren) {
-  const store = useStore();
-  const authState = useAtomValue(authStateAtom);
-  const refreshAuth = useRefreshAuth();
-
-  useEffect(() => {
-    configureApiClient({
-      getAccessToken: () => store.get(authStateAtom).token,
-      refreshAccessToken: async () => {
-        const result = await refreshAuth();
-        return result.token;
-      },
-    });
-  }, [refreshAuth, store]);
-
-  useEffect(() => {
-    refreshAuth().catch(() => {
-      // Refresh handles auth state changes for expected auth failures.
-    });
-  }, [refreshAuth]);
-
-  useEffect(() => {
-    if (authState.status !== 'authenticated' || !authState.token) return;
-
-    let timeout: ReturnType<typeof setTimeout> | undefined;
-    let disposed = false;
-    let refreshing = false;
-
-    const clearRefreshTimer = () => {
-      if (timeout !== undefined) {
-        clearTimeout(timeout);
-        timeout = undefined;
-      }
-    };
-
-    const scheduleRefresh = (delayMs: number) => {
-      clearRefreshTimer();
-      timeout = setTimeout(runRefresh, Math.max(0, delayMs));
-    };
-
-    const retryIfStillDue = () => {
-      const current = store.get(authStateAtom);
-      if (current.status !== 'authenticated' || !current.token) return;
-
-      const delay = getAuthRefreshDelayMs(current.token);
-      if (delay !== undefined && delay <= 0) {
-        scheduleRefresh(REFRESH_RETRY_DELAY_MS);
-      }
-    };
-
-    function runRefresh() {
-      if (disposed || refreshing) return;
-      refreshing = true;
-      clearRefreshTimer();
-      refreshAuth()
-        .catch(() => {
-          // 401s move auth state to guest. Other failures should retry later.
-        })
-        .finally(() => {
-          refreshing = false;
-          if (!disposed) retryIfStillDue();
-        });
-    }
-
-    const refreshIfDue = () => {
-      const current = store.get(authStateAtom);
-      if (current.status !== 'authenticated' || !current.token) return;
-
-      const delay = getAuthRefreshDelayMs(current.token);
-      if (delay !== undefined && delay <= 0) {
-        runRefresh();
-      }
-    };
-
-    const refreshIfVisible = () => {
-      if (document.visibilityState === 'visible') {
-        refreshIfDue();
-      }
-    };
-
-    const delay = getAuthRefreshDelayMs(authState.token);
-    if (delay !== undefined) {
-      scheduleRefresh(delay);
-    }
-
-    window.addEventListener('focus', refreshIfDue);
-    window.addEventListener('online', refreshIfDue);
-    document.addEventListener('visibilitychange', refreshIfVisible);
-
-    return () => {
-      disposed = true;
-      clearRefreshTimer();
-      window.removeEventListener('focus', refreshIfDue);
-      window.removeEventListener('online', refreshIfDue);
-      document.removeEventListener('visibilitychange', refreshIfVisible);
-    };
-  }, [authState.status, authState.token, refreshAuth, store]);
-
-  return children;
 }
