--- libs/client/shell/src/runtime/auth-stub.ts
@@ -1,3 +1,0 @@
-import type {AuthStateValue} from '@shipfox/client-auth';
-
-export const authStub: AuthStateValue | undefined = undefined;
