+++ libs/client/shell/src/runtime/provider-stack.tsx
@@ -4,6 +4,7 @@ import {type QueryClient, QueryClientProvider} from '@tanstack/react-query';
 import {type createStore, Provider as JotaiProvider} from 'jotai';
 import type {PropsWithChildren, ReactNode} from 'react';
 import type {ClientFeature} from '#contract.js';
+import {AuthRuntime, type AuthRuntimeProps} from './auth.js';
 
 type Store = ReturnType<typeof createStore>;
 
@@ -11,11 +12,13 @@ export function ShellProviderStack({
   features,
   queryClient,
   store,
+  auth,
   children,
 }: PropsWithChildren<{
   features: readonly ClientFeature[];
   queryClient: QueryClient;
   store: Store;
+  auth?: Pick<AuthRuntimeProps, 'effects'>;
 }>) {
   const featureProviders = features.flatMap((feature) => feature.providers ?? []);
   const nestedProviders = featureProviders.reduceRight<ReactNode>(
@@ -26,7 +29,9 @@ export function ShellProviderStack({
     <ThemeProvider>
       <TooltipProvider>
         <QueryClientProvider client={queryClient}>
-          <JotaiProvider store={store}>{nestedProviders}</JotaiProvider>
+          <JotaiProvider store={store}>
+            <AuthRuntime {...auth}>{nestedProviders}</AuthRuntime>
+          </JotaiProvider>
         </QueryClientProvider>
       </TooltipProvider>
     </ThemeProvider>
