+++ libs/client/shell/src/runtime/providers.test.tsx
@@ -31,7 +31,12 @@ describe('ShellProviderStack', () => {
     ];
 
     render(
-      <ShellProviderStack features={features} queryClient={queryClient} store={store}>
+      <ShellProviderStack
+        features={features}
+        queryClient={queryClient}
+        store={store}
+        auth={{effects: false}}
+      >
         <div>Content</div>
       </ShellProviderStack>,
     );
